This commit revamps the Projects section.
Author had a big list of projects of which only a few projects were
featured, and the rest were displayed in a grid. This commit removes
that project display and instead modifies the featured.js as per the
current authors taste.

Later when we have a big list of projects we can revert this change for
a better UI.